Cast any i32 to u32 in order to account for differences between Windows and Linux

+++ readstat/tests/get_var_names_test.rs
@@ -14,7 +14,7 @@ fn get_var_names() {
     let mut d = readstat::ReadStatData::new(rsp);
     let error = d.get_metadata().unwrap();
 
-    assert_eq!(error, readstat_sys::readstat_error_e_READSTAT_OK);
+    assert_eq!(error, readstat_sys::readstat_error_e_READSTAT_OK as u32);
 
     let vars = d.vars;
